refactor: 파싱에러문제해결
- reviewAverage Int -> Double로 변경

+++ mmtalk/Data/DTO/ProductListDTO.swift
@@ -43,7 +43,7 @@ struct ProductListDTO: APIResponse {
         let sellPrice: Int
         let imageURL: String
         let reviewCount: Int
-        let reviewAverage: Int
+        let reviewAverage: Double
         let tags: [Tag]
         let discountRate: Int
         private let partnerNo: Int
